Fixes bug in Mongo install on centos 7
The Mongo install on Centos 7 has bug whereby the
stop function in /etc/init.d/mongodb expects the dbpath
entry in /etc/mongodb.conf to have spaces around the = sign.
Given that hasn't been fixed, this changes our elements to put
the spaces in the conf file so stop works.

Change-Id: I9aee65facfdac6cedf3882c86213c5e0df2a0eb7
Closes-bug: DBAAS-238