Do not remove '/etc/my.cnf'
The Problem:
Trove expects certain directory structure.
For instance, it places 'my.cnf' and overrides to
'/etc/mysql/' and '/etc/mysql/conf.d/' on CentOS and Debian.
Not all datastores (MariaDB 10.0) however search those directory.
Depending on how it is compiled and packaged, MariaDB 10.0
may be looking only for '/etc/my.cnf'.

The solution:
We create '/etc/my.cnf' if it does not exist
and append includes to '/etc/mysql' and '/etc/mysql/conf.d/'
to it. We then copy this file to '/etc/mysql/my.cnf'
where Trove will be looking for it. We keep
the original file with include paths in '/etc' so that
all datastores find any Trove created files.

Change-Id: I0796f5b3e2e69b453f4eb8f42458901c3dba9e6c
Closes-bug: DBAAS-272 and DBAAS-129