Workaround for multiple devstacks on network
Redstack, by default, configures devstack to operate on a private
network 10.0.0.0/24 with eth0 connected directly to br100, causing
layer 2 ARP packets to be broadcast to all VMs and guest VMs
inside them. So if two devstacks have guest VMs with local ip
10.0.0.2, they will both send back ARP reponses, with different
MAC addresses. This can cause intermittent unresponsiveness.
This temporary hack makes it unlikely that guest instances will
have collisions with guests running in other devstacks within
the same ARP broadcast domain, by setting the second octet of the
devstack private network to be the last octet of the address of eth0

Change-Id: I4d4c55f98db811c416e5ae470cf62c2221a9411c
Closes-Bug: dbaas-505