Allow specifying devstack branch in redstack
At present, redstack clones the most recent version of devstack. Sometimes this
is undesirable as devstack can at times become unstable. A solution to this
would be to add an environment variable or option to redstack to specify the
branch of devstack to use (with the default remaining 'master').

With this change you can now specify:

DEVSTACK_BRANCH=stable/kilo ./redstack install

and redstack would switch to the stable/kilo branch before running './stack.sh'
Optionally, for persistence, it can be defined in either options.rc or
~/.redstack.options.rc and it will be picked up from there.

The branch is switched, even if ithe devstack directory already existed.

If an invalid branch is specified, redstack stops with an error message.

Change-Id: I50c64db917bb87bac5d1b1d4d0231270577e26a8
Closes-Bug: #1449627

Set branch when clone other projects

When cloning other projects we need to set the branch
according to the DEVSTACK_BRANCH env variable
Currently hacked it since a couple of the projects
don't have stable/kilo branches in which case they
stay on master
This can get cleaned up later

Change-Id: Id169aceb935bbe119aba384aed5092619f32fb25