Create a volatile directory for Cassandra PID
Cassandra on RHEL-based systems needs a volatile directory
for its PID files. The '/var/run/' where the PIDs are stored
is mounted on a temporary filesystem. The process directories
therefore have to be recreated on every boot.
The operating system has a facility for this.
A configuration file with folder definition has to be placed to
'/lib/tmpfiles.d'.

Change-Id: Id2dc962d1eb314a07e08a6676c133ff19f94e6e2
Closes-Bug: DBAAS-275