Fixes MariaDB failure due to missing pidfile option
MariaDB installs without any options specified in my.cnf. This
causes the pid-file on initial boot to be named for the current
hostname. At the time of mysql start, in certain cases, the hostname
hasn't be reset and is named for the IP addr. Later in the boot
sequence the guest hostname is changed and mysql stop will no
longer work because it can't find the pid file. This causes trove
create to fail.

Change-Id: Ib5e87e8c2a467c772ba19ed2bce1f4f8cfe3654c