PostgreSQL guest on Centos fails to start
It turns out that the Centos elements did not get all fixes that
the RHEL elements got.
The only difference between RHEL and Centos elements is the used
YUM repository (and the location of THP setting).
I broke the scripts into pre-install and install parts.
The install part is shared between the two platforms.

Change-Id: I082afc9ac347cd0bcb950f695f865fd895d59a6c
Closes-Bug: DBAAS-975