Pin oslo.context to v2.1 or less
The new version of oslo.context (v2.2.0) introduced a
roles parameter that causes an issue when a guest with
that version communicates to a conductor with the older
version. This will always happen for non-master branches
because the core requirements file has oslo.context pinned
to an older version but the guest requirements file didn't.

Change-Id: Ib244f2278ad9c99ba3f8d916b48f101674842d98
Closes-bug: dbaas-1540