Change service type, ExecStart and ExecStop in dbora.service
Simplied the dbora.service by changing the ExecStart and ExecStop to
call dbstart and dbshut directly, deprecated the startup.sh and
shutdown.sh scripts.

Also changed the service type from simple to forking, because Oracle
should actually be forked and run as a daemon when systemd starts it
up.

Change-Id: Ic292108adcab29462fe9e76d3629e3fd183404b1
Partial-Bug: dbaas-1575