Add 'fault' flavor to test persisting an error msg
In order to test that error message are now persisted,
a method of guaranteeing that a Trove create command
will fail is needed.  The simplest way to do that is to
attempt to allocate more resources from Nova than it
has available.  To this effect, a 'fault' flavor has
been added.

Change-Id: I8af04bcabc46d3a4f75e371cdab60055f2b8df95