Fix Couchbase 3.0 backup failure on RHEL
CB3.0 backup is failing on RHEL/Centos due to issues with the
install/configuration of the Snappy compression library. It turns
out that CB3.0 isn't really supported on Centos7/RHEL7 so we need
to build the C++ portion of the library from source and use pip
to install the python portion.

Change-Id: I775ec1df80d69d6934f599082de76b2449d276eb
Closes-bug: dbaas-1760