(fix gate) Don't source openrc in gate run
This is a manual cherry pick of Iec71474ccfe757c8384c0b4e4b0758636668aebe

The latest set of changes to devstack include one where enabling a
plugin more than once now causes a fatal error. Trovestack code
sources openrc, which sources other files that eventually call
enable_plugin on the trove module. Unfortunately since trovestack also
sources some of the same files, this causes the error to be
triggered. The fix is to remove the call to openrc and just set the
credentials manually.

Change-Id: I887c980c75a093ce9c76bcd032b07069d61336ce
Co-Authored-By: Peter Stachowski <peter@tesora.com>