Add support for Postgresql 9.6
Added elements to for pgsql v9.6. As version 9.5+ includes
pg_rewind, we no longer have to pull source and build it.
The code for this was kept in the elements but only used if
VERSION=9.4.

Redstack was refactored a bit to accomodate to handle kick-starting
multiple versions of a datastore.

Do not push this upstream - the upstream commit is [1]

[1] https://review.openstack.org/409524

Change-Id: If85e418ff3bf1a5db9a4f91ca484c8c5ec63ce84
Closes-bug: dbaas-2432